Don't use a Gargoyle-internal function
gli_strict_warning is internal to Gargoyle, and is not meant for
external use. I'm planning on switching Gargoyle to C++, which will
result in garglk.h only working with C++ (it will expose an empty file,
effectively, to C code).

Since gli_strict_warning simply dumps the message to stderr, I've
created a JACL-specific function to do the same thing here, so the
behavior should be the same.