Convert to char* for strlen
strlen expects a char*, not unsigned char*. Passing unsigned char* is
technically a constraint violation (compilers are free to reject it),
though in practice it will result in a warning. But the warning is
correct, and this cast silences it.

A better approach may be to switch things around so that char* is used
throughout, instead of casting, but that's a more involved change.