Prefer _WIN32 to WIN32
_WIN32 is defined by the compiler and should exist whenever JACL is
built on Windows. I'm no Windows expert, but it seems that WIN32 is an
oddity. It's defined by some Windows headers, which means it won't exist
till windows.h (or something else) is included. As it stands, it seems
that JACL expects WIN32 to be defined by the build system when on
Windows.

My rationale here is that a Windows compiler already knows it's on
Windows, making the need to tell it so redundant. Things will be less
fragile if _WIN32 is checked, since it requires no user intervention.

This also completely removes a couple lines for Windows that write
"\r\n" instead of "\n" at the end of a line: the file being written to
is opened in text mode (using "w", not "wb"). In text mode, newline
translation is already done by the C library. Writing "\n" will cause
the Windows stdio implementation to write out "\r\n" (and likewise, a
"\r\n" when read is converted to "\n"). So the manual writing of CR is
unnecessary, even on Windows.